% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{calculateInterest}
\alias{calculateInterest}
\title{Calculate Interest}
\usage{
calculateInterest(current_balance, APR)
}
\arguments{
\item{current_balance}{current ending balance for the specific debt for the month}

\item{APR}{Yearly rate of interest}
}
\value{
amount of interest that would be added to the total amount of debt for this month
}
\description{
Calculates the interest added to the balance given the debt ending balance and its calculated monthly rate.
}
