% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{calculatePaidBalance}
\alias{calculatePaidBalance}
\title{Title Calculate Paid Balance}
\usage{
calculatePaidBalance(tackle, original_balance, disposable_income, minimum)
}
\arguments{
\item{tackle}{boolean value that signals if this debt will use disposible income to tackle the debt faster}

\item{original_balance}{starting balance for the debt in this month}

\item{disposable_income}{amount of disposible income available to this debt}

\item{minimum}{The minimum payment that must be paid to this card}
}
\value{
a data frame with columns: balance, the updated debt balance / residual: amount of disposable income left over after this debt / debtWiped: boolean value to know if debt has been completely zeroed
}
\description{
Calculates the updated new balance for a certain debt after monthly payment depending if we will use extra money to tackle the debt or if we will just pay a minimum payment
}
