% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{calculateTotalBalance}
\alias{calculateTotalBalance}
\title{Calculate the Total Balance}
\usage{
calculateTotalBalance(current_month, result_df)
}
\arguments{
\item{current_month}{month we want to calculate the total debt balance}

\item{result_df}{Data frame containing all month's debt balance informations}
}
\value{
the amount of total debt balance owed throughout all debt accounts
}
\description{
Calculates the total balance owed from all different debts
}
