% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{checkColumnClasses}
\alias{checkColumnClasses}
\title{checkColumnClasses}
\usage{
checkColumnClasses(df, expected_classes, columns_required, function_name)
}
\arguments{
\item{df}{data frame we will validate}

\item{expected_classes}{expected classes for the columns we require}

\item{columns_required}{columns in the dataframw we want to check}

\item{function_name}{Name of the function we are checking for debugging}
}
\value{
True if there is no issue, error if there is an error
}
\description{
validates column classes to catch unexpected casts
}
