% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{createTitleSection}
\alias{createTitleSection}
\title{Create Title Section}
\usage{
createTitleSection(
  heading_text,
  background_color,
  box_content,
  font_color,
  padding,
  box_height,
  margin_bottom = "20px"
)
}
\arguments{
\item{heading_text}{Content you want to add to the H3 title}

\item{box_content}{Content you want to add below the H3 title
All other params are pretty self explanatory}
}
\value{
title section into the UI
}
\description{
Creates multiple value boxes when given their id titles
}
