% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{createValueBox}
\alias{createValueBox}
\title{CreateValueBox}
\usage{
createValueBox(title, value)
}
\arguments{
\item{title}{information detailing what the number above represents}

\item{value}{The variable value we want to display to the user about their debt data throughout the months}
}
\value{
Creates a single value box
}
\description{
Where you create how the value boxes look and where they display information. It can be expanded later to become more general and have different templates.
}
