% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{findPreviousBalance}
\alias{findPreviousBalance}
\title{Find Previous Balance}
\usage{
findPreviousBalance(month_index, current_title, result_df, column_retrival)
}
\arguments{
\item{month_index}{Current or consequent month for which we want to get the previous balance for}

\item{result_df}{Data frame where we can find the data with all previous month information stored}

\item{column_retrival}{Name of column we want to retrieve from}

\item{title}{The title of the debt we want to retrieve information for}
}
\value{
Generally it is the previous value of the requested debt from the requested column. This will be usually the previous month balance.
}
\description{
Used to retreive a certain debt's previous month value. Typically used for the balance but made allow any column retrival
}
