% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{multipleValueBoxes}
\alias{multipleValueBoxes}
\title{MultipleValueBoxes}
\usage{
multipleValueBoxes(id_titles)
}
\arguments{
\item{id_titles}{a vector containing the chosen names for each value box unit you want to create in the ui}
}
\value{
create x amount of value boxes to the ui
}
\description{
Creates multiple value boxes when given their id titles
}
