% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{noChangeSimulation}
\alias{noChangeSimulation}
\title{No Change Simulation}
\usage{
noChangeSimulation(debt_df)
}
\arguments{
\item{debt_df}{data frame containing all basic debt information collected from the user. (title, balance, APR, minimum payment)}
}
\value{
a data frame with a row for each debt's progress per month for total_months amount of months.
}
\description{
Creates a simulation of just paying minimum payments for the debt information given in the amount of month's they could be debt free using disposible income
}
