% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{simulateProgress}
\alias{simulateProgress}
\title{Title Simulate Progress}
\usage{
simulateProgress(debt_df, disposable_df)
}
\arguments{
\item{debt_df}{Data frame containing all initial debt information such as debt title, balance, APR and minimum}

\item{disposable_df}{Data frame that contains the disposible income information the user will have available after their budgetted needs}
}
\value{
A data frame with a row for each debt's progress per month. With columns ....
}
\description{
Simulates the payoff journey of all the debts submitted through the UI if they were to use their disposable money each month to implement the snowball method
}
