% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{sumMinimums}
\alias{sumMinimums}
\title{SumMinimums}
\usage{
sumMinimums(debt_df, index)
}
\arguments{
\item{debt_df}{Data frame with the collected debt data with the users that contains minimum}

\item{index}{integer index that signifies the latest paid off debt}
}
\value{
a float number representating the total minimum payment total for the debts that have been paid off
}
\description{
SumMinimums
}
