% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{updateInputUnit}
\alias{updateInputUnit}
\title{Update Input Unit}
\usage{
updateInputUnit(session, ...)
}
\arguments{
\item{session}{a Shiny server built in parameter that keeps track of the session is needed to be passed}

\item{...}{optional amounts of list with ID information for the UI item you will want to update}
}
\description{
This function resets the value of the input unit to become blank. Its main use is to reset the values after a debt or disposible income have been submitted
}
